add script to download benchmark text
add bench/getdata.py, a horrid python script to obtain three text files.
used tqdm to show progress bars. there was some decompressing and
concatenating of files, which could have been inoffensively written
in bash. because of the progress bar logic I used python, which was
a mess. some of these files are multiple gigabytes, so not showing
any progress bar looks like the program is hanging, but this is not
the answer.